moved helper function tests
